+++ src/runtime_src/xdp/profile/device/device_intf.h
@@ -48,7 +48,7 @@ namespace xdp {
 // Helper methods
 
 XDP_EXPORT
-uint32_t GetDeviceTraceBufferSize(uint32_t property);
+uint64_t GetDeviceTraceBufferSize(uint32_t property);
 
 XDP_EXPORT
 uint64_t GetTS2MMBufSize(bool isAIETrace = false);
